Relocate verifier repair intent parse/validate/apply pipeline
Phase 2 batch 9 for Issue #682. Moves the 12 fns + 3 supporting
struct types (~340 LOC of bodies) for the verifier repair intent
parse -> validate -> apply -> post-apply pipeline out of turn.rs
into verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  Parse (4, #[cfg(test)]):
    - parse_verifier_repair_intent_reply
    - parse_verifier_repair_intents_reply
    - parse_verifier_repair_patch_proposal_reply
    - patch_proposal_to_verifier_repair_intents

  Validation entry points (3):
    - validate_verifier_repair_intent (#[cfg(test)])
    - validate_verifier_repair_intents (#[cfg(test)])
    - validate_verifier_repair_intents_with_accepted_plan (production)

  Validation pipeline core (5):
    - load_repair_validation_target
    - prepare_verifier_repair_edit_payloads
    - apply_verifier_repair_edit_payloads
    - validate_verifier_repair_test_constraints
    - validate_verifier_repair_post_apply_candidate
    - validate_verifier_repair_intents_inner (orchestrator)
    - build_repair_test_import_contract_evidence

  Supporting struct types (3):
    - RepairValidationTargetState
    - PreparedRepairIntentEdits
    - AppliedRepairCandidate

All promoted to pub(super); the 6 originally-#[cfg(test)] fns retain
the gate in verifier_orchestration.rs.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::repair_driver::VERIFIER_REPAIR_PASS_MAX_EDIT_BYTES
    - super::repair_patch_validation::{ValidatedVerifierRepairEdit,
      ValidationFailure, VerifierRepairIntent,
      validate_accepted_repair_plan_authorizes_target} (production
      block, plus removed the now-duplicate VerifierRepairIntent
      from the #[cfg(test)] sub-import)
    - crate::util::file_classify::is_test_file
    - std::path::PathBuf

  * turn.rs:
    - Extended the production use super::verifier_orchestration::{}
      with validate_verifier_repair_intents_with_accepted_plan only
      (the other 4 validate/parse fns moved to the #[cfg(test)] block).
    - Added #[cfg(test)] use super::repair_patch_validation::
      VerifierRepairIntent; so the in-mod test fixtures that build
      VerifierRepairIntent literals (~20 sites) still resolve at
      the turn:: namespace, and dropped VerifierRepairIntent +
      is_test_file + VERIFIER_REPAIR_PASS_MAX_EDIT_BYTES +
      VERIFIER_REPAIR_PASS_MAX_EDITS from the production imports.
    - Added the 4 cfg(test) fns to the existing
      `use super::super::verifier_orchestration::{...}` block in
      mod progress_tests.
    - Added VerifierRepairIntent to mod progress_tests
      `use super::{...}` block so unqualified literals resolve.
    - Cleaned 6 orphaned #[cfg(test)] attribute lines left over
      the deleted fn sites (clippy
      empty_line_after_doc_comments under -D warnings).

Metrics: turn.rs 33,668 -> 33,333 (-335);
verifier_orchestration.rs 1,380 -> 1,724 (+344). Cumulative Issue
#682 progress: turn.rs 34,958 -> 33,333 (-1,625, ~54% of Phase 2
target). Acceptance gap: 33,333 -> 32,000 = ~1,333 LOC remaining.
